requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
onnxruntime==1.19.2
torch==2.5.0
transformers==4.45.2